Token ID validation, local storage removed, view for password reset page

+++ Web_App/Rest/Favorite/Controller/FavoriteController.cs
@@ -1,8 +1,10 @@
 ﻿using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using System.Diagnostics.Contracts;
+using System.Drawing;
 using Web_App.Rest.Favorite.Models;
 using Web_App.Rest.Favorite.Services;
+using Web_App.Rest.JWT.Services;
 
 namespace Web_App.Rest.Favorite.Controller
 {
@@ -11,9 +13,11 @@ namespace Web_App.Rest.Favorite.Controller
     public class FavoriteController : ControllerBase
     {
         FavoriteService _favoriteService;
-        public FavoriteController()
+        private readonly ITokenService _tokenService;
+        public FavoriteController(IConfiguration _conf)
         {
             _favoriteService = new FavoriteService();
+            _tokenService = new TokenService(_conf);
         }
 
         [Authorize]
@@ -21,6 +25,11 @@ namespace Web_App.Rest.Favorite.Controller
         [Route("addfavorite")]
         public IActionResult AddFavorite([FromBody] FavoriteRequestModel model)
         {
+            string token = Request.Cookies["AccessToken"];
+            if (!_tokenService.IDQueryTokenVerificator(model.ClientID, token))
+            {
+                return BadRequest(new { message = "UnAuthorized Attempt to Access Data belong to Other User!" });
+            }
             _favoriteService.addToFavorite(model);
             return Ok(new {message = "Added successfully" });
         }
@@ -30,6 +39,11 @@ namespace Web_App.Rest.Favorite.Controller
         [Route("getlistfavorite")]
         public IActionResult GetIndexFavorite([FromQuery] int clientId)
         {
+            string token = Request.Cookies["AccessToken"];
+            if (!_tokenService.IDQueryTokenVerificator(clientId, token))
+            {
+                return BadRequest(new { message = "UnAuthorized Attempt to Access Data belong to Other User!" });
+            }
             List<int> indexes = new List<int>();
             indexes = _favoriteService.getAllIndexesFavoriteClient(clientId);
             return Ok(new { ListIndex = indexes });
@@ -40,6 +54,11 @@ namespace Web_App.Rest.Favorite.Controller
         [Route("removefavoriteitem")]
         public IActionResult RemoveFavoriteItem([FromBody] FavoriteRequestModel model)
         {
+            string token = Request.Cookies["AccessToken"];
+            if (!_tokenService.IDQueryTokenVerificator(model.ClientID, token))
+            {
+                return BadRequest(new { message = "UnAuthorized Attempt to Access Data belong to Other User!" });
+            }
             _favoriteService.removeFovorite(model);
             return Ok(new { message = "Deleted successfully" });
         }
@@ -49,6 +68,11 @@ namespace Web_App.Rest.Favorite.Controller
         [Route("getallfavoriteuser")]
         public IActionResult GetAllFavoriteItem(int userID)
         {
+            string token = Request.Cookies["AccessToken"];
+            if (!_tokenService.IDQueryTokenVerificator(userID, token))
+            {
+                return BadRequest(new { message = "UnAuthorized Attempt to Access Data belong to Other User!" });
+            }
             List<FavoritePageResponseModel> response = new List<FavoritePageResponseModel>();
             response = _favoriteService.getAllUserFavorite(userID);
             float sum = response.Sum(t => t.Cost);
