+++ Web_App/Rest/Order/Controller/OrderController.cs
@@ -1,6 +1,7 @@
 ﻿using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using System.Collections.Generic;
+using Web_App.Rest.JWT.Services;
 using Web_App.Rest.Order.Model;
 using Web_App.Rest.Order.Service;
 
@@ -11,17 +12,23 @@ namespace Web_App.Rest.Order.Controller
     [Authorize]
     public class OrderController : ControllerBase
     {
-
+        private readonly ITokenService _tokenService;
         private OrderService _orderService;
-        public OrderController()
+        public OrderController(IConfiguration _conf)
         {
             _orderService = new OrderService();
+            _tokenService = new TokenService(_conf);
         }
 
         [HttpPost]
         [Route("addorder")]
         public IActionResult addOrder(OrderRequestModel orderRequestModel)
         {
+            string token = Request.Cookies["AccessToken"];
+            if (!_tokenService.IDQueryTokenVerificator(orderRequestModel.ClientID, token))
+            {
+                return BadRequest(new { message = "UnAuthorized Attempt to Access Data belong to Other User!" });
+            }
             _orderService.addOrderToDB(orderRequestModel);
             _orderService.addOrderProductToTable(orderRequestModel.TowarCount);
             _orderService.removeProductFromCart();
@@ -32,6 +39,11 @@ namespace Web_App.Rest.Order.Controller
         [HttpGet]
         public IActionResult getOrderDetails([FromQuery] int orderID, int clientID)
         {
+            string token = Request.Cookies["AccessToken"];
+            if (!_tokenService.IDQueryTokenVerificator(clientID, token))
+            {
+                return BadRequest(new { message = "UnAuthorized Attempt to Access Data belong to Other User!" });
+            }
             OrderDetailsModel orderDetailsModel = new OrderDetailsModel();
             orderDetailsModel = _orderService.getOrderDetailsModel(orderID, clientID);
             return Ok(orderDetailsModel);
@@ -40,6 +52,11 @@ namespace Web_App.Rest.Order.Controller
         [Route("getallordersuser")]
         public IActionResult getOrdersUser([FromQuery] int  userID)
         {
+            string token = Request.Cookies["AccessToken"];
+            if (!_tokenService.IDQueryTokenVerificator(userID, token))
+            {
+                return BadRequest(new { message = "UnAuthorized Attempt to Access Data belong to Other User!" });
+            }
             List <OrdersUserModel> model = new List <OrdersUserModel>();
             model = _orderService.getOrdersUsers(userID);
